Output a dataframe to a table in RTF with opinionated formatting.

df_to_rtf is a convenience wrapper around r2rtf functionality for outputting an RTF table, and optionally with at least one colspan in the column headers.

Usage

df_to_rtf(
  df,
  file_name,
  org,
  header_page,
  table_title,
  header_table0,
  header_table1,
  col_widths0,
  col_widths1,
  col_just0,
  col_just1,
  border_top1,
  footnote_table,
  cell_indents,
  footer_snapshot,
  path_program,
  path_output,
  font_family = 6,
  font_size = 10,
  font_size_header = 10,
  col_header_bg_color = "gray94",
  orientation = "landscape",
  margin = c(1, 1, 0.75, 0.75, 0.5, 0.5),
  title_as_heading = 2,
  bookmark = TRUE,
  bookmark_inc_title = FALSE,
  output = TRUE
)

Arguments

df

data.frame. Dataframe to be output to RTF. Required.

file_name

character. File path where the RTF file will be saved. Required. Strongly suggested to not include spaces in the file name to avoid issues with the bookmark functionality.

org

character. Organization name, placed in the top line of the page header, left-aligned on the same line as the right-aligned page number. Default is "" (no organization name).

header_page

character. Header text for the page, typically the name of the study.

table_title

character. Title of the table, the table caption. Required.

header_table0

character vector. Column spanning header for the table. Length of vector must be less than number of columns in df. Optional.

header_table1

character vector. Column headers for the table. Length of vector must equal number of columns in df. Required.

col_widths0

numeric vector. Relative widths for the column spanning header. Optional. Only used if header_table0 is entered. The total sum of col_widths0 must equal the total sum of col_widths1.

col_widths1

numeric vector. Relative widths for the column headers.

col_just0

character vector. Justification for the column spanning header. Optional. Only used if header_table0 is entered.

col_just1

character vector. Justification for the column headers.

border_top1

character vector. Border style for the top border of each column header. Optional. Length of vector must equal number of columns in df.

footnote_table

character vector. Footnotes for the table, if any.

cell_indents

numeric vector. Indentation for each cell in the table, in twips. Default is 0, no indentation. cell_indents must be a vector of length nrow(df)*ncol(df). The first value controls the indentation of the cell at row 1, column 1; the second value controls the indentation of the cell at row 1, column 2; and so on.

footer_snapshot

character. Data snapshot used, placed in top row of the page footer. Required.

path_program

character. Path to the program that generated the table, e.g., "path/to/program.R", placed in second row of the footer. Required.

path_output

character. Path to the output file, e.g., "path/to/output.rtf", placed in third row of the footer. If not entered, will use file_name.

font_family

numeric. Font family for the RTF document. Default is 6 (Calibri). See r2rtf:::font_type() for other options.

font_size

numeric. Font size for the body of the table, including table text, column headers, and table footnote. Default is 10.

font_size_header

numeric. Font size for the header and footer text. Default is 10.

col_header_bg_color

character. Background color for the column header rows, default is "gray94".

orientation

character. Orientation of the page, either "portrait" or "landscape". Default is "landscape".

margin

numeric vector. The value sets left, right, top, bottom, header and footer margin in order. Default is c(1, 1, 0.75, 0.75, 0.5, 0.5).

title_as_heading

numeric. If non-missing, the table title will be inserted as a Word heading style. Expected input is a value of 1, 2, 3, or 4. Default is 2. If no title style desired, set to 0 or do not specify in the function call.

bookmark

logical. If TRUE, a bookmark will be created in the RTF file with the same name as the file (without extension).

bookmark_inc_title

logical. If TRUE, the bookmark will be placed in the RTF file such that the title of the table is included. Otherwise, the title is excluded. Default is FALSE.

output

logical. If TRUE, the RTF file will be written to the location specified in file_name. Otherwise, the RTF object is returned. Default is TRUE.

Value

An RTF file, file_name.

Details

df_to_rtf creates a formatted RTF file from a data frame input with optional colspans in the column headers. See example for how to specify colspans.

Examples

df <- data.frame(Col1="1", Col2="2", Col3="3", Col4="4", Col5="5")

# Example: simple table without colspans
df_to_rtf(df              = df,
          file_name       = "MyTable.rtf", 
          table_title     = "Table 1. My table",
          col_widths1     = c(3, 2, 2, 2, 2),
          col_just1       = c("l", "c", "c", "c", "c"),
          footer_snapshot = "Analysis 2025", 
          path_program    = "C:\\\\program\\\\path\\\\code.r",
          path_output     = "C:\\\\output\\\\path\\\\MyTable.rtf",
          output          = FALSE)
#> $start
#> [1] "{\\rtf1\\ansi\n\\deff0\\deflang1033\n{\\fonttbl{\\f0\\froman\\fcharset1\\fprq2 Times New Roman;}\n{\\f1\\froman\\fcharset161\\fprq2 Times New Roman Greek;}\n{\\f2\\fswiss\\fcharset161\\fprq2 Arial Greek;}\n{\\f3\\fswiss\\fcharset0\\fprq2 Arial;}\n{\\f4\\fswiss\\fcharset1\\fprq2 Helvetica;}\n{\\f5\\fswiss\\fcharset1\\fprq2 Calibri;}\n{\\f6\\froman\\fcharset1\\fprq2 Georgia;}\n{\\f7\\ffroman\\fcharset1\\fprq2 Cambria;}\n{\\f8\\fmodern\\fcharset0\\fprq2 Courier New;}\n{\\f9\\ftech\\fcharset2\\fprq2 Symbol;}\n}\n\n{\\colortbl\n;\n\\red255\\green255\\blue255;\n\\red240\\green248\\blue255;\n\\red250\\green235\\blue215;\n\\red255\\green239\\blue219;\n\\red238\\green223\\blue204;\n\\red205\\green192\\blue176;\n\\red139\\green131\\blue120;\n\\red127\\green255\\blue212;\n\\red127\\green255\\blue212;\n\\red118\\green238\\blue198;\n\\red102\\green205\\blue170;\n\\red69\\green139\\blue116;\n\\red240\\green255\\blue255;\n\\red240\\green255\\blue255;\n\\red224\\green238\\blue238;\n\\red193\\green205\\blue205;\n\\red131\\green139\\blue139;\n\\red245\\green245\\blue220;\n\\red255\\green228\\blue196;\n\\red255\\green228\\blue196;\n\\red238\\green213\\blue183;\n\\red205\\green183\\blue158;\n\\red139\\green125\\blue107;\n\\red0\\green0\\blue0;\n\\red255\\green235\\blue205;\n\\red0\\green0\\blue255;\n\\red0\\green0\\blue255;\n\\red0\\green0\\blue238;\n\\red0\\green0\\blue205;\n\\red0\\green0\\blue139;\n\\red138\\green43\\blue226;\n\\red165\\green42\\blue42;\n\\red255\\green64\\blue64;\n\\red238\\green59\\blue59;\n\\red205\\green51\\blue51;\n\\red139\\green35\\blue35;\n\\red222\\green184\\blue135;\n\\red255\\green211\\blue155;\n\\red238\\green197\\blue145;\n\\red205\\green170\\blue125;\n\\red139\\green115\\blue85;\n\\red95\\green158\\blue160;\n\\red152\\green245\\blue255;\n\\red142\\green229\\blue238;\n\\red122\\green197\\blue205;\n\\red83\\green134\\blue139;\n\\red127\\green255\\blue0;\n\\red127\\green255\\blue0;\n\\red118\\green238\\blue0;\n\\red102\\green205\\blue0;\n\\red69\\green139\\blue0;\n\\red210\\green105\\blue30;\n\\red255\\green127\\blue36;\n\\red238\\green118\\blue33;\n\\red205\\green102\\blue29;\n\\red139\\green69\\blue19;\n\\red255\\green127\\blue80;\n\\red255\\green114\\blue86;\n\\red238\\green106\\blue80;\n\\red205\\green91\\blue69;\n\\red139\\green62\\blue47;\n\\red100\\green149\\blue237;\n\\red255\\green248\\blue220;\n\\red255\\green248\\blue220;\n\\red238\\green232\\blue205;\n\\red205\\green200\\blue177;\n\\red139\\green136\\blue120;\n\\red0\\green255\\blue255;\n\\red0\\green255\\blue255;\n\\red0\\green238\\blue238;\n\\red0\\green205\\blue205;\n\\red0\\green139\\blue139;\n\\red0\\green0\\blue139;\n\\red0\\green139\\blue139;\n\\red184\\green134\\blue11;\n\\red255\\green185\\blue15;\n\\red238\\green173\\blue14;\n\\red205\\green149\\blue12;\n\\red139\\green101\\blue8;\n\\red169\\green169\\blue169;\n\\red0\\green100\\blue0;\n\\red169\\green169\\blue169;\n\\red189\\green183\\blue107;\n\\red139\\green0\\blue139;\n\\red85\\green107\\blue47;\n\\red202\\green255\\blue112;\n\\red188\\green238\\blue104;\n\\red162\\green205\\blue90;\n\\red110\\green139\\blue61;\n\\red255\\green140\\blue0;\n\\red255\\green127\\blue0;\n\\red238\\green118\\blue0;\n\\red205\\green102\\blue0;\n\\red139\\green69\\blue0;\n\\red153\\green50\\blue204;\n\\red191\\green62\\blue255;\n\\red178\\green58\\blue238;\n\\red154\\green50\\blue205;\n\\red104\\green34\\blue139;\n\\red139\\green0\\blue0;\n\\red233\\green150\\blue122;\n\\red143\\green188\\blue143;\n\\red193\\green255\\blue193;\n\\red180\\green238\\blue180;\n\\red155\\green205\\blue155;\n\\red105\\green139\\blue105;\n\\red72\\green61\\blue139;\n\\red47\\green79\\blue79;\n\\red151\\green255\\blue255;\n\\red141\\green238\\blue238;\n\\red121\\green205\\blue205;\n\\red82\\green139\\blue139;\n\\red47\\green79\\blue79;\n\\red0\\green206\\blue209;\n\\red148\\green0\\blue211;\n\\red255\\green20\\blue147;\n\\red255\\green20\\blue147;\n\\red238\\green18\\blue137;\n\\red205\\green16\\blue118;\n\\red139\\green10\\blue80;\n\\red0\\green191\\blue255;\n\\red0\\green191\\blue255;\n\\red0\\green178\\blue238;\n\\red0\\green154\\blue205;\n\\red0\\green104\\blue139;\n\\red105\\green105\\blue105;\n\\red105\\green105\\blue105;\n\\red30\\green144\\blue255;\n\\red30\\green144\\blue255;\n\\red28\\green134\\blue238;\n\\red24\\green116\\blue205;\n\\red16\\green78\\blue139;\n\\red178\\green34\\blue34;\n\\red255\\green48\\blue48;\n\\red238\\green44\\blue44;\n\\red205\\green38\\blue38;\n\\red139\\green26\\blue26;\n\\red255\\green250\\blue240;\n\\red34\\green139\\blue34;\n\\red220\\green220\\blue220;\n\\red248\\green248\\blue255;\n\\red255\\green215\\blue0;\n\\red255\\green215\\blue0;\n\\red238\\green201\\blue0;\n\\red205\\green173\\blue0;\n\\red139\\green117\\blue0;\n\\red218\\green165\\blue32;\n\\red255\\green193\\blue37;\n\\red238\\green180\\blue34;\n\\red205\\green155\\blue29;\n\\red139\\green105\\blue20;\n\\red190\\green190\\blue190;\n\\red0\\green0\\blue0;\n\\red3\\green3\\blue3;\n\\red5\\green5\\blue5;\n\\red8\\green8\\blue8;\n\\red10\\green10\\blue10;\n\\red13\\green13\\blue13;\n\\red15\\green15\\blue15;\n\\red18\\green18\\blue18;\n\\red20\\green20\\blue20;\n\\red23\\green23\\blue23;\n\\red26\\green26\\blue26;\n\\red28\\green28\\blue28;\n\\red31\\green31\\blue31;\n\\red33\\green33\\blue33;\n\\red36\\green36\\blue36;\n\\red38\\green38\\blue38;\n\\red41\\green41\\blue41;\n\\red43\\green43\\blue43;\n\\red46\\green46\\blue46;\n\\red48\\green48\\blue48;\n\\red51\\green51\\blue51;\n\\red54\\green54\\blue54;\n\\red56\\green56\\blue56;\n\\red59\\green59\\blue59;\n\\red61\\green61\\blue61;\n\\red64\\green64\\blue64;\n\\red66\\green66\\blue66;\n\\red69\\green69\\blue69;\n\\red71\\green71\\blue71;\n\\red74\\green74\\blue74;\n\\red77\\green77\\blue77;\n\\red79\\green79\\blue79;\n\\red82\\green82\\blue82;\n\\red84\\green84\\blue84;\n\\red87\\green87\\blue87;\n\\red89\\green89\\blue89;\n\\red92\\green92\\blue92;\n\\red94\\green94\\blue94;\n\\red97\\green97\\blue97;\n\\red99\\green99\\blue99;\n\\red102\\green102\\blue102;\n\\red105\\green105\\blue105;\n\\red107\\green107\\blue107;\n\\red110\\green110\\blue110;\n\\red112\\green112\\blue112;\n\\red115\\green115\\blue115;\n\\red117\\green117\\blue117;\n\\red120\\green120\\blue120;\n\\red122\\green122\\blue122;\n\\red125\\green125\\blue125;\n\\red127\\green127\\blue127;\n\\red130\\green130\\blue130;\n\\red133\\green133\\blue133;\n\\red135\\green135\\blue135;\n\\red138\\green138\\blue138;\n\\red140\\green140\\blue140;\n\\red143\\green143\\blue143;\n\\red145\\green145\\blue145;\n\\red148\\green148\\blue148;\n\\red150\\green150\\blue150;\n\\red153\\green153\\blue153;\n\\red156\\green156\\blue156;\n\\red158\\green158\\blue158;\n\\red161\\green161\\blue161;\n\\red163\\green163\\blue163;\n\\red166\\green166\\blue166;\n\\red168\\green168\\blue168;\n\\red171\\green171\\blue171;\n\\red173\\green173\\blue173;\n\\red176\\green176\\blue176;\n\\red179\\green179\\blue179;\n\\red181\\green181\\blue181;\n\\red184\\green184\\blue184;\n\\red186\\green186\\blue186;\n\\red189\\green189\\blue189;\n\\red191\\green191\\blue191;\n\\red194\\green194\\blue194;\n\\red196\\green196\\blue196;\n\\red199\\green199\\blue199;\n\\red201\\green201\\blue201;\n\\red204\\green204\\blue204;\n\\red207\\green207\\blue207;\n\\red209\\green209\\blue209;\n\\red212\\green212\\blue212;\n\\red214\\green214\\blue214;\n\\red217\\green217\\blue217;\n\\red219\\green219\\blue219;\n\\red222\\green222\\blue222;\n\\red224\\green224\\blue224;\n\\red227\\green227\\blue227;\n\\red229\\green229\\blue229;\n\\red232\\green232\\blue232;\n\\red235\\green235\\blue235;\n\\red237\\green237\\blue237;\n\\red240\\green240\\blue240;\n\\red242\\green242\\blue242;\n\\red245\\green245\\blue245;\n\\red247\\green247\\blue247;\n\\red250\\green250\\blue250;\n\\red252\\green252\\blue252;\n\\red255\\green255\\blue255;\n\\red0\\green255\\blue0;\n\\red0\\green255\\blue0;\n\\red0\\green238\\blue0;\n\\red0\\green205\\blue0;\n\\red0\\green139\\blue0;\n\\red173\\green255\\blue47;\n\\red190\\green190\\blue190;\n\\red0\\green0\\blue0;\n\\red3\\green3\\blue3;\n\\red5\\green5\\blue5;\n\\red8\\green8\\blue8;\n\\red10\\green10\\blue10;\n\\red13\\green13\\blue13;\n\\red15\\green15\\blue15;\n\\red18\\green18\\blue18;\n\\red20\\green20\\blue20;\n\\red23\\green23\\blue23;\n\\red26\\green26\\blue26;\n\\red28\\green28\\blue28;\n\\red31\\green31\\blue31;\n\\red33\\green33\\blue33;\n\\red36\\green36\\blue36;\n\\red38\\green38\\blue38;\n\\red41\\green41\\blue41;\n\\red43\\green43\\blue43;\n\\red46\\green46\\blue46;\n\\red48\\green48\\blue48;\n\\red51\\green51\\blue51;\n\\red54\\green54\\blue54;\n\\red56\\green56\\blue56;\n\\red59\\green59\\blue59;\n\\red61\\green61\\blue61;\n\\red64\\green64\\blue64;\n\\red66\\green66\\blue66;\n\\red69\\green69\\blue69;\n\\red71\\green71\\blue71;\n\\red74\\green74\\blue74;\n\\red77\\green77\\blue77;\n\\red79\\green79\\blue79;\n\\red82\\green82\\blue82;\n\\red84\\green84\\blue84;\n\\red87\\green87\\blue87;\n\\red89\\green89\\blue89;\n\\red92\\green92\\blue92;\n\\red94\\green94\\blue94;\n\\red97\\green97\\blue97;\n\\red99\\green99\\blue99;\n\\red102\\green102\\blue102;\n\\red105\\green105\\blue105;\n\\red107\\green107\\blue107;\n\\red110\\green110\\blue110;\n\\red112\\green112\\blue112;\n\\red115\\green115\\blue115;\n\\red117\\green117\\blue117;\n\\red120\\green120\\blue120;\n\\red122\\green122\\blue122;\n\\red125\\green125\\blue125;\n\\red127\\green127\\blue127;\n\\red130\\green130\\blue130;\n\\red133\\green133\\blue133;\n\\red135\\green135\\blue135;\n\\red138\\green138\\blue138;\n\\red140\\green140\\blue140;\n\\red143\\green143\\blue143;\n\\red145\\green145\\blue145;\n\\red148\\green148\\blue148;\n\\red150\\green150\\blue150;\n\\red153\\green153\\blue153;\n\\red156\\green156\\blue156;\n\\red158\\green158\\blue158;\n\\red161\\green161\\blue161;\n\\red163\\green163\\blue163;\n\\red166\\green166\\blue166;\n\\red168\\green168\\blue168;\n\\red171\\green171\\blue171;\n\\red173\\green173\\blue173;\n\\red176\\green176\\blue176;\n\\red179\\green179\\blue179;\n\\red181\\green181\\blue181;\n\\red184\\green184\\blue184;\n\\red186\\green186\\blue186;\n\\red189\\green189\\blue189;\n\\red191\\green191\\blue191;\n\\red194\\green194\\blue194;\n\\red196\\green196\\blue196;\n\\red199\\green199\\blue199;\n\\red201\\green201\\blue201;\n\\red204\\green204\\blue204;\n\\red207\\green207\\blue207;\n\\red209\\green209\\blue209;\n\\red212\\green212\\blue212;\n\\red214\\green214\\blue214;\n\\red217\\green217\\blue217;\n\\red219\\green219\\blue219;\n\\red222\\green222\\blue222;\n\\red224\\green224\\blue224;\n\\red227\\green227\\blue227;\n\\red229\\green229\\blue229;\n\\red232\\green232\\blue232;\n\\red235\\green235\\blue235;\n\\red237\\green237\\blue237;\n\\red240\\green240\\blue240;\n\\red242\\green242\\blue242;\n\\red245\\green245\\blue245;\n\\red247\\green247\\blue247;\n\\red250\\green250\\blue250;\n\\red252\\green252\\blue252;\n\\red255\\green255\\blue255;\n\\red240\\green255\\blue240;\n\\red240\\green255\\blue240;\n\\red224\\green238\\blue224;\n\\red193\\green205\\blue193;\n\\red131\\green139\\blue131;\n\\red255\\green105\\blue180;\n\\red255\\green110\\blue180;\n\\red238\\green106\\blue167;\n\\red205\\green96\\blue144;\n\\red139\\green58\\blue98;\n\\red205\\green92\\blue92;\n\\red255\\green106\\blue106;\n\\red238\\green99\\blue99;\n\\red205\\green85\\blue85;\n\\red139\\green58\\blue58;\n\\red255\\green255\\blue240;\n\\red255\\green255\\blue240;\n\\red238\\green238\\blue224;\n\\red205\\green205\\blue193;\n\\red139\\green139\\blue131;\n\\red240\\green230\\blue140;\n\\red255\\green246\\blue143;\n\\red238\\green230\\blue133;\n\\red205\\green198\\blue115;\n\\red139\\green134\\blue78;\n\\red230\\green230\\blue250;\n\\red255\\green240\\blue245;\n\\red255\\green240\\blue245;\n\\red238\\green224\\blue229;\n\\red205\\green193\\blue197;\n\\red139\\green131\\blue134;\n\\red124\\green252\\blue0;\n\\red255\\green250\\blue205;\n\\red255\\green250\\blue205;\n\\red238\\green233\\blue191;\n\\red205\\green201\\blue165;\n\\red139\\green137\\blue112;\n\\red173\\green216\\blue230;\n\\red191\\green239\\blue255;\n\\red178\\green223\\blue238;\n\\red154\\green192\\blue205;\n\\red104\\green131\\blue139;\n\\red240\\green128\\blue128;\n\\red224\\green255\\blue255;\n\\red224\\green255\\blue255;\n\\red209\\green238\\blue238;\n\\red180\\green205\\blue205;\n\\red122\\green139\\blue139;\n\\red238\\green221\\blue130;\n\\red255\\green236\\blue139;\n\\red238\\green220\\blue130;\n\\red205\\green190\\blue112;\n\\red139\\green129\\blue76;\n\\red250\\green250\\blue210;\n\\red211\\green211\\blue211;\n\\red144\\green238\\blue144;\n\\red211\\green211\\blue211;\n\\red255\\green182\\blue193;\n\\red255\\green174\\blue185;\n\\red238\\green162\\blue173;\n\\red205\\green140\\blue149;\n\\red139\\green95\\blue101;\n\\red255\\green160\\blue122;\n\\red255\\green160\\blue122;\n\\red238\\green149\\blue114;\n\\red205\\green129\\blue98;\n\\red139\\green87\\blue66;\n\\red32\\green178\\blue170;\n\\red135\\green206\\blue250;\n\\red176\\green226\\blue255;\n\\red164\\green211\\blue238;\n\\red141\\green182\\blue205;\n\\red96\\green123\\blue139;\n\\red132\\green112\\blue255;\n\\red119\\green136\\blue153;\n\\red119\\green136\\blue153;\n\\red176\\green196\\blue222;\n\\red202\\green225\\blue255;\n\\red188\\green210\\blue238;\n\\red162\\green181\\blue205;\n\\red110\\green123\\blue139;\n\\red255\\green255\\blue224;\n\\red255\\green255\\blue224;\n\\red238\\green238\\blue209;\n\\red205\\green205\\blue180;\n\\red139\\green139\\blue122;\n\\red50\\green205\\blue50;\n\\red250\\green240\\blue230;\n\\red255\\green0\\blue255;\n\\red255\\green0\\blue255;\n\\red238\\green0\\blue238;\n\\red205\\green0\\blue205;\n\\red139\\green0\\blue139;\n\\red176\\green48\\blue96;\n\\red255\\green52\\blue179;\n\\red238\\green48\\blue167;\n\\red205\\green41\\blue144;\n\\red139\\green28\\blue98;\n\\red102\\green205\\blue170;\n\\red0\\green0\\blue205;\n\\red186\\green85\\blue211;\n\\red224\\green102\\blue255;\n\\red209\\green95\\blue238;\n\\red180\\green82\\blue205;\n\\red122\\green55\\blue139;\n\\red147\\green112\\blue219;\n\\red171\\green130\\blue255;\n\\red159\\green121\\blue238;\n\\red137\\green104\\blue205;\n\\red93\\green71\\blue139;\n\\red60\\green179\\blue113;\n\\red123\\green104\\blue238;\n\\red0\\green250\\blue154;\n\\red72\\green209\\blue204;\n\\red199\\green21\\blue133;\n\\red25\\green25\\blue112;\n\\red245\\green255\\blue250;\n\\red255\\green228\\blue225;\n\\red255\\green228\\blue225;\n\\red238\\green213\\blue210;\n\\red205\\green183\\blue181;\n\\red139\\green125\\blue123;\n\\red255\\green228\\blue181;\n\\red255\\green222\\blue173;\n\\red255\\green222\\blue173;\n\\red238\\green207\\blue161;\n\\red205\\green179\\blue139;\n\\red139\\green121\\blue94;\n\\red0\\green0\\blue128;\n\\red0\\green0\\blue128;\n\\red253\\green245\\blue230;\n\\red107\\green142\\blue35;\n\\red192\\green255\\blue62;\n\\red179\\green238\\blue58;\n\\red154\\green205\\blue50;\n\\red105\\green139\\blue34;\n\\red255\\green165\\blue0;\n\\red255\\green165\\blue0;\n\\red238\\green154\\blue0;\n\\red205\\green133\\blue0;\n\\red139\\green90\\blue0;\n\\red255\\green69\\blue0;\n\\red255\\green69\\blue0;\n\\red238\\green64\\blue0;\n\\red205\\green55\\blue0;\n\\red139\\green37\\blue0;\n\\red218\\green112\\blue214;\n\\red255\\green131\\blue250;\n\\red238\\green122\\blue233;\n\\red205\\green105\\blue201;\n\\red139\\green71\\blue137;\n\\red238\\green232\\blue170;\n\\red152\\green251\\blue152;\n\\red154\\green255\\blue154;\n\\red144\\green238\\blue144;\n\\red124\\green205\\blue124;\n\\red84\\green139\\blue84;\n\\red175\\green238\\blue238;\n\\red187\\green255\\blue255;\n\\red174\\green238\\blue238;\n\\red150\\green205\\blue205;\n\\red102\\green139\\blue139;\n\\red219\\green112\\blue147;\n\\red255\\green130\\blue171;\n\\red238\\green121\\blue159;\n\\red205\\green104\\blue137;\n\\red139\\green71\\blue93;\n\\red255\\green239\\blue213;\n\\red255\\green218\\blue185;\n\\red255\\green218\\blue185;\n\\red238\\green203\\blue173;\n\\red205\\green175\\blue149;\n\\red139\\green119\\blue101;\n\\red205\\green133\\blue63;\n\\red255\\green192\\blue203;\n\\red255\\green181\\blue197;\n\\red238\\green169\\blue184;\n\\red205\\green145\\blue158;\n\\red139\\green99\\blue108;\n\\red221\\green160\\blue221;\n\\red255\\green187\\blue255;\n\\red238\\green174\\blue238;\n\\red205\\green150\\blue205;\n\\red139\\green102\\blue139;\n\\red176\\green224\\blue230;\n\\red160\\green32\\blue240;\n\\red155\\green48\\blue255;\n\\red145\\green44\\blue238;\n\\red125\\green38\\blue205;\n\\red85\\green26\\blue139;\n\\red255\\green0\\blue0;\n\\red255\\green0\\blue0;\n\\red238\\green0\\blue0;\n\\red205\\green0\\blue0;\n\\red139\\green0\\blue0;\n\\red188\\green143\\blue143;\n\\red255\\green193\\blue193;\n\\red238\\green180\\blue180;\n\\red205\\green155\\blue155;\n\\red139\\green105\\blue105;\n\\red65\\green105\\blue225;\n\\red72\\green118\\blue255;\n\\red67\\green110\\blue238;\n\\red58\\green95\\blue205;\n\\red39\\green64\\blue139;\n\\red139\\green69\\blue19;\n\\red250\\green128\\blue114;\n\\red255\\green140\\blue105;\n\\red238\\green130\\blue98;\n\\red205\\green112\\blue84;\n\\red139\\green76\\blue57;\n\\red244\\green164\\blue96;\n\\red46\\green139\\blue87;\n\\red84\\green255\\blue159;\n\\red78\\green238\\blue148;\n\\red67\\green205\\blue128;\n\\red46\\green139\\blue87;\n\\red255\\green245\\blue238;\n\\red255\\green245\\blue238;\n\\red238\\green229\\blue222;\n\\red205\\green197\\blue191;\n\\red139\\green134\\blue130;\n\\red160\\green82\\blue45;\n\\red255\\green130\\blue71;\n\\red238\\green121\\blue66;\n\\red205\\green104\\blue57;\n\\red139\\green71\\blue38;\n\\red135\\green206\\blue235;\n\\red135\\green206\\blue255;\n\\red126\\green192\\blue238;\n\\red108\\green166\\blue205;\n\\red74\\green112\\blue139;\n\\red106\\green90\\blue205;\n\\red131\\green111\\blue255;\n\\red122\\green103\\blue238;\n\\red105\\green89\\blue205;\n\\red71\\green60\\blue139;\n\\red112\\green128\\blue144;\n\\red198\\green226\\blue255;\n\\red185\\green211\\blue238;\n\\red159\\green182\\blue205;\n\\red108\\green123\\blue139;\n\\red112\\green128\\blue144;\n\\red255\\green250\\blue250;\n\\red255\\green250\\blue250;\n\\red238\\green233\\blue233;\n\\red205\\green201\\blue201;\n\\red139\\green137\\blue137;\n\\red0\\green255\\blue127;\n\\red0\\green255\\blue127;\n\\red0\\green238\\blue118;\n\\red0\\green205\\blue102;\n\\red0\\green139\\blue69;\n\\red70\\green130\\blue180;\n\\red99\\green184\\blue255;\n\\red92\\green172\\blue238;\n\\red79\\green148\\blue205;\n\\red54\\green100\\blue139;\n\\red210\\green180\\blue140;\n\\red255\\green165\\blue79;\n\\red238\\green154\\blue73;\n\\red205\\green133\\blue63;\n\\red139\\green90\\blue43;\n\\red216\\green191\\blue216;\n\\red255\\green225\\blue255;\n\\red238\\green210\\blue238;\n\\red205\\green181\\blue205;\n\\red139\\green123\\blue139;\n\\red255\\green99\\blue71;\n\\red255\\green99\\blue71;\n\\red238\\green92\\blue66;\n\\red205\\green79\\blue57;\n\\red139\\green54\\blue38;\n\\red64\\green224\\blue208;\n\\red0\\green245\\blue255;\n\\red0\\green229\\blue238;\n\\red0\\green197\\blue205;\n\\red0\\green134\\blue139;\n\\red238\\green130\\blue238;\n\\red208\\green32\\blue144;\n\\red255\\green62\\blue150;\n\\red238\\green58\\blue140;\n\\red205\\green50\\blue120;\n\\red139\\green34\\blue82;\n\\red245\\green222\\blue179;\n\\red255\\green231\\blue186;\n\\red238\\green216\\blue174;\n\\red205\\green186\\blue150;\n\\red139\\green126\\blue102;\n\\red245\\green245\\blue245;\n\\red255\\green255\\blue0;\n\\red255\\green255\\blue0;\n\\red238\\green238\\blue0;\n\\red205\\green205\\blue0;\n\\red139\\green139\\blue0;\n\\red154\\green205\\blue50;\n} \n{\\stylesheet{\\s1 heading 1;}\n{\\s2 heading 2;}\n{\\s3 heading 3;}\n{\\s4 heading 4;}\n}\n\n"
#> 
#> $body
#> [1] "{\\header\n{\\pard\\tqr\\tx12960\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b  \\tab Page \\chpgn  of {\\field{\\*\\fldinst NUMPAGES }} }\\par}\n{\\pard\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b }\\par}\n}\n{\\footer\n{\\pard\\tqr\\tx12960\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b DB lock (snapshot): Analysis 2025 \\tab Created on: 14APR2026 00:24:34 UTC}\\par}\n{\\pard\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b Program name: C:\\\\program\\\\path\\\\code.r}\\par}\n{\\pard\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b Output name: C:\\\\output\\\\path\\\\MyTable.rtf}\\par}\n}\n\\paperw15840\\paperh12240\\landscape\n\n\\margl1440\\margr1440\\margt1080\\margb1080\\headery720\\footery720\n\n{\\pard\\hyphpar\\sb180\\sa180\\fi0\\li0\\ri0\\ql\\fs24\\s2{\\f5\\b Table 1. My table}\\par}\n\n\n{\\*\\bkmkstart MyTable}\\trowd\\trgaph108\\trleft0\\trqc\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx3535\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx5891\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx8247\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx10603\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrr\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx12959\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 Col1}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 Col2}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 Col3}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 Col4}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 Col5}\\cell\n\\intbl\\row\\pard\n\\trowd\\trgaph108\\trleft0\\trqc\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrs\\brdrw15\\clvertalt\\cellx3535\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrs\\brdrw15\\clvertalt\\cellx5891\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrs\\brdrw15\\clvertalt\\cellx8247\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrs\\brdrw15\\clvertalt\\cellx10603\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrr\\brdrs\\brdrw15\\clbrdrb\\brdrs\\brdrw15\\clvertalt\\cellx12959\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5 1}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5 2}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5 3}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5 4}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5 5}\\cell\n\\intbl\\row\\pard\n\n\n"
#> 
#> $end
#> [1] "{\\*\\bkmkend MyTable}\n}"
#> 

# Example: table with colspans
# - Cspan1 spans Col2 and Col3, Cspan2 spans Col4 and Col5
df_to_rtf(df              = df,
          file_name       = "MyColSpanTable.rtf", 
          table_title     = "Table 1. My table",
          header_table0   = c(" ", "Cspan1", "CSpan2"), 
          col_widths0     = c(3, 4, 4),
          col_widths1     = c(3, 2, 2, 2, 2),
          col_just0       = c("l", "c", "c"), 
          col_just1       = c("l", "c", "c", "c", "c"),
          footer_snapshot = "Analysis 2025", 
          path_program    = "C:\\\\program\\\\path\\\\code.r",
          path_output     = "C:\\\\output\\\\path\\\\MyColSpanTable.rtf",
          output          = FALSE)
#> $start
#> [1] "{\\rtf1\\ansi\n\\deff0\\deflang1033\n{\\fonttbl{\\f0\\froman\\fcharset1\\fprq2 Times New Roman;}\n{\\f1\\froman\\fcharset161\\fprq2 Times New Roman Greek;}\n{\\f2\\fswiss\\fcharset161\\fprq2 Arial Greek;}\n{\\f3\\fswiss\\fcharset0\\fprq2 Arial;}\n{\\f4\\fswiss\\fcharset1\\fprq2 Helvetica;}\n{\\f5\\fswiss\\fcharset1\\fprq2 Calibri;}\n{\\f6\\froman\\fcharset1\\fprq2 Georgia;}\n{\\f7\\ffroman\\fcharset1\\fprq2 Cambria;}\n{\\f8\\fmodern\\fcharset0\\fprq2 Courier New;}\n{\\f9\\ftech\\fcharset2\\fprq2 Symbol;}\n}\n\n{\\colortbl\n;\n\\red255\\green255\\blue255;\n\\red240\\green248\\blue255;\n\\red250\\green235\\blue215;\n\\red255\\green239\\blue219;\n\\red238\\green223\\blue204;\n\\red205\\green192\\blue176;\n\\red139\\green131\\blue120;\n\\red127\\green255\\blue212;\n\\red127\\green255\\blue212;\n\\red118\\green238\\blue198;\n\\red102\\green205\\blue170;\n\\red69\\green139\\blue116;\n\\red240\\green255\\blue255;\n\\red240\\green255\\blue255;\n\\red224\\green238\\blue238;\n\\red193\\green205\\blue205;\n\\red131\\green139\\blue139;\n\\red245\\green245\\blue220;\n\\red255\\green228\\blue196;\n\\red255\\green228\\blue196;\n\\red238\\green213\\blue183;\n\\red205\\green183\\blue158;\n\\red139\\green125\\blue107;\n\\red0\\green0\\blue0;\n\\red255\\green235\\blue205;\n\\red0\\green0\\blue255;\n\\red0\\green0\\blue255;\n\\red0\\green0\\blue238;\n\\red0\\green0\\blue205;\n\\red0\\green0\\blue139;\n\\red138\\green43\\blue226;\n\\red165\\green42\\blue42;\n\\red255\\green64\\blue64;\n\\red238\\green59\\blue59;\n\\red205\\green51\\blue51;\n\\red139\\green35\\blue35;\n\\red222\\green184\\blue135;\n\\red255\\green211\\blue155;\n\\red238\\green197\\blue145;\n\\red205\\green170\\blue125;\n\\red139\\green115\\blue85;\n\\red95\\green158\\blue160;\n\\red152\\green245\\blue255;\n\\red142\\green229\\blue238;\n\\red122\\green197\\blue205;\n\\red83\\green134\\blue139;\n\\red127\\green255\\blue0;\n\\red127\\green255\\blue0;\n\\red118\\green238\\blue0;\n\\red102\\green205\\blue0;\n\\red69\\green139\\blue0;\n\\red210\\green105\\blue30;\n\\red255\\green127\\blue36;\n\\red238\\green118\\blue33;\n\\red205\\green102\\blue29;\n\\red139\\green69\\blue19;\n\\red255\\green127\\blue80;\n\\red255\\green114\\blue86;\n\\red238\\green106\\blue80;\n\\red205\\green91\\blue69;\n\\red139\\green62\\blue47;\n\\red100\\green149\\blue237;\n\\red255\\green248\\blue220;\n\\red255\\green248\\blue220;\n\\red238\\green232\\blue205;\n\\red205\\green200\\blue177;\n\\red139\\green136\\blue120;\n\\red0\\green255\\blue255;\n\\red0\\green255\\blue255;\n\\red0\\green238\\blue238;\n\\red0\\green205\\blue205;\n\\red0\\green139\\blue139;\n\\red0\\green0\\blue139;\n\\red0\\green139\\blue139;\n\\red184\\green134\\blue11;\n\\red255\\green185\\blue15;\n\\red238\\green173\\blue14;\n\\red205\\green149\\blue12;\n\\red139\\green101\\blue8;\n\\red169\\green169\\blue169;\n\\red0\\green100\\blue0;\n\\red169\\green169\\blue169;\n\\red189\\green183\\blue107;\n\\red139\\green0\\blue139;\n\\red85\\green107\\blue47;\n\\red202\\green255\\blue112;\n\\red188\\green238\\blue104;\n\\red162\\green205\\blue90;\n\\red110\\green139\\blue61;\n\\red255\\green140\\blue0;\n\\red255\\green127\\blue0;\n\\red238\\green118\\blue0;\n\\red205\\green102\\blue0;\n\\red139\\green69\\blue0;\n\\red153\\green50\\blue204;\n\\red191\\green62\\blue255;\n\\red178\\green58\\blue238;\n\\red154\\green50\\blue205;\n\\red104\\green34\\blue139;\n\\red139\\green0\\blue0;\n\\red233\\green150\\blue122;\n\\red143\\green188\\blue143;\n\\red193\\green255\\blue193;\n\\red180\\green238\\blue180;\n\\red155\\green205\\blue155;\n\\red105\\green139\\blue105;\n\\red72\\green61\\blue139;\n\\red47\\green79\\blue79;\n\\red151\\green255\\blue255;\n\\red141\\green238\\blue238;\n\\red121\\green205\\blue205;\n\\red82\\green139\\blue139;\n\\red47\\green79\\blue79;\n\\red0\\green206\\blue209;\n\\red148\\green0\\blue211;\n\\red255\\green20\\blue147;\n\\red255\\green20\\blue147;\n\\red238\\green18\\blue137;\n\\red205\\green16\\blue118;\n\\red139\\green10\\blue80;\n\\red0\\green191\\blue255;\n\\red0\\green191\\blue255;\n\\red0\\green178\\blue238;\n\\red0\\green154\\blue205;\n\\red0\\green104\\blue139;\n\\red105\\green105\\blue105;\n\\red105\\green105\\blue105;\n\\red30\\green144\\blue255;\n\\red30\\green144\\blue255;\n\\red28\\green134\\blue238;\n\\red24\\green116\\blue205;\n\\red16\\green78\\blue139;\n\\red178\\green34\\blue34;\n\\red255\\green48\\blue48;\n\\red238\\green44\\blue44;\n\\red205\\green38\\blue38;\n\\red139\\green26\\blue26;\n\\red255\\green250\\blue240;\n\\red34\\green139\\blue34;\n\\red220\\green220\\blue220;\n\\red248\\green248\\blue255;\n\\red255\\green215\\blue0;\n\\red255\\green215\\blue0;\n\\red238\\green201\\blue0;\n\\red205\\green173\\blue0;\n\\red139\\green117\\blue0;\n\\red218\\green165\\blue32;\n\\red255\\green193\\blue37;\n\\red238\\green180\\blue34;\n\\red205\\green155\\blue29;\n\\red139\\green105\\blue20;\n\\red190\\green190\\blue190;\n\\red0\\green0\\blue0;\n\\red3\\green3\\blue3;\n\\red5\\green5\\blue5;\n\\red8\\green8\\blue8;\n\\red10\\green10\\blue10;\n\\red13\\green13\\blue13;\n\\red15\\green15\\blue15;\n\\red18\\green18\\blue18;\n\\red20\\green20\\blue20;\n\\red23\\green23\\blue23;\n\\red26\\green26\\blue26;\n\\red28\\green28\\blue28;\n\\red31\\green31\\blue31;\n\\red33\\green33\\blue33;\n\\red36\\green36\\blue36;\n\\red38\\green38\\blue38;\n\\red41\\green41\\blue41;\n\\red43\\green43\\blue43;\n\\red46\\green46\\blue46;\n\\red48\\green48\\blue48;\n\\red51\\green51\\blue51;\n\\red54\\green54\\blue54;\n\\red56\\green56\\blue56;\n\\red59\\green59\\blue59;\n\\red61\\green61\\blue61;\n\\red64\\green64\\blue64;\n\\red66\\green66\\blue66;\n\\red69\\green69\\blue69;\n\\red71\\green71\\blue71;\n\\red74\\green74\\blue74;\n\\red77\\green77\\blue77;\n\\red79\\green79\\blue79;\n\\red82\\green82\\blue82;\n\\red84\\green84\\blue84;\n\\red87\\green87\\blue87;\n\\red89\\green89\\blue89;\n\\red92\\green92\\blue92;\n\\red94\\green94\\blue94;\n\\red97\\green97\\blue97;\n\\red99\\green99\\blue99;\n\\red102\\green102\\blue102;\n\\red105\\green105\\blue105;\n\\red107\\green107\\blue107;\n\\red110\\green110\\blue110;\n\\red112\\green112\\blue112;\n\\red115\\green115\\blue115;\n\\red117\\green117\\blue117;\n\\red120\\green120\\blue120;\n\\red122\\green122\\blue122;\n\\red125\\green125\\blue125;\n\\red127\\green127\\blue127;\n\\red130\\green130\\blue130;\n\\red133\\green133\\blue133;\n\\red135\\green135\\blue135;\n\\red138\\green138\\blue138;\n\\red140\\green140\\blue140;\n\\red143\\green143\\blue143;\n\\red145\\green145\\blue145;\n\\red148\\green148\\blue148;\n\\red150\\green150\\blue150;\n\\red153\\green153\\blue153;\n\\red156\\green156\\blue156;\n\\red158\\green158\\blue158;\n\\red161\\green161\\blue161;\n\\red163\\green163\\blue163;\n\\red166\\green166\\blue166;\n\\red168\\green168\\blue168;\n\\red171\\green171\\blue171;\n\\red173\\green173\\blue173;\n\\red176\\green176\\blue176;\n\\red179\\green179\\blue179;\n\\red181\\green181\\blue181;\n\\red184\\green184\\blue184;\n\\red186\\green186\\blue186;\n\\red189\\green189\\blue189;\n\\red191\\green191\\blue191;\n\\red194\\green194\\blue194;\n\\red196\\green196\\blue196;\n\\red199\\green199\\blue199;\n\\red201\\green201\\blue201;\n\\red204\\green204\\blue204;\n\\red207\\green207\\blue207;\n\\red209\\green209\\blue209;\n\\red212\\green212\\blue212;\n\\red214\\green214\\blue214;\n\\red217\\green217\\blue217;\n\\red219\\green219\\blue219;\n\\red222\\green222\\blue222;\n\\red224\\green224\\blue224;\n\\red227\\green227\\blue227;\n\\red229\\green229\\blue229;\n\\red232\\green232\\blue232;\n\\red235\\green235\\blue235;\n\\red237\\green237\\blue237;\n\\red240\\green240\\blue240;\n\\red242\\green242\\blue242;\n\\red245\\green245\\blue245;\n\\red247\\green247\\blue247;\n\\red250\\green250\\blue250;\n\\red252\\green252\\blue252;\n\\red255\\green255\\blue255;\n\\red0\\green255\\blue0;\n\\red0\\green255\\blue0;\n\\red0\\green238\\blue0;\n\\red0\\green205\\blue0;\n\\red0\\green139\\blue0;\n\\red173\\green255\\blue47;\n\\red190\\green190\\blue190;\n\\red0\\green0\\blue0;\n\\red3\\green3\\blue3;\n\\red5\\green5\\blue5;\n\\red8\\green8\\blue8;\n\\red10\\green10\\blue10;\n\\red13\\green13\\blue13;\n\\red15\\green15\\blue15;\n\\red18\\green18\\blue18;\n\\red20\\green20\\blue20;\n\\red23\\green23\\blue23;\n\\red26\\green26\\blue26;\n\\red28\\green28\\blue28;\n\\red31\\green31\\blue31;\n\\red33\\green33\\blue33;\n\\red36\\green36\\blue36;\n\\red38\\green38\\blue38;\n\\red41\\green41\\blue41;\n\\red43\\green43\\blue43;\n\\red46\\green46\\blue46;\n\\red48\\green48\\blue48;\n\\red51\\green51\\blue51;\n\\red54\\green54\\blue54;\n\\red56\\green56\\blue56;\n\\red59\\green59\\blue59;\n\\red61\\green61\\blue61;\n\\red64\\green64\\blue64;\n\\red66\\green66\\blue66;\n\\red69\\green69\\blue69;\n\\red71\\green71\\blue71;\n\\red74\\green74\\blue74;\n\\red77\\green77\\blue77;\n\\red79\\green79\\blue79;\n\\red82\\green82\\blue82;\n\\red84\\green84\\blue84;\n\\red87\\green87\\blue87;\n\\red89\\green89\\blue89;\n\\red92\\green92\\blue92;\n\\red94\\green94\\blue94;\n\\red97\\green97\\blue97;\n\\red99\\green99\\blue99;\n\\red102\\green102\\blue102;\n\\red105\\green105\\blue105;\n\\red107\\green107\\blue107;\n\\red110\\green110\\blue110;\n\\red112\\green112\\blue112;\n\\red115\\green115\\blue115;\n\\red117\\green117\\blue117;\n\\red120\\green120\\blue120;\n\\red122\\green122\\blue122;\n\\red125\\green125\\blue125;\n\\red127\\green127\\blue127;\n\\red130\\green130\\blue130;\n\\red133\\green133\\blue133;\n\\red135\\green135\\blue135;\n\\red138\\green138\\blue138;\n\\red140\\green140\\blue140;\n\\red143\\green143\\blue143;\n\\red145\\green145\\blue145;\n\\red148\\green148\\blue148;\n\\red150\\green150\\blue150;\n\\red153\\green153\\blue153;\n\\red156\\green156\\blue156;\n\\red158\\green158\\blue158;\n\\red161\\green161\\blue161;\n\\red163\\green163\\blue163;\n\\red166\\green166\\blue166;\n\\red168\\green168\\blue168;\n\\red171\\green171\\blue171;\n\\red173\\green173\\blue173;\n\\red176\\green176\\blue176;\n\\red179\\green179\\blue179;\n\\red181\\green181\\blue181;\n\\red184\\green184\\blue184;\n\\red186\\green186\\blue186;\n\\red189\\green189\\blue189;\n\\red191\\green191\\blue191;\n\\red194\\green194\\blue194;\n\\red196\\green196\\blue196;\n\\red199\\green199\\blue199;\n\\red201\\green201\\blue201;\n\\red204\\green204\\blue204;\n\\red207\\green207\\blue207;\n\\red209\\green209\\blue209;\n\\red212\\green212\\blue212;\n\\red214\\green214\\blue214;\n\\red217\\green217\\blue217;\n\\red219\\green219\\blue219;\n\\red222\\green222\\blue222;\n\\red224\\green224\\blue224;\n\\red227\\green227\\blue227;\n\\red229\\green229\\blue229;\n\\red232\\green232\\blue232;\n\\red235\\green235\\blue235;\n\\red237\\green237\\blue237;\n\\red240\\green240\\blue240;\n\\red242\\green242\\blue242;\n\\red245\\green245\\blue245;\n\\red247\\green247\\blue247;\n\\red250\\green250\\blue250;\n\\red252\\green252\\blue252;\n\\red255\\green255\\blue255;\n\\red240\\green255\\blue240;\n\\red240\\green255\\blue240;\n\\red224\\green238\\blue224;\n\\red193\\green205\\blue193;\n\\red131\\green139\\blue131;\n\\red255\\green105\\blue180;\n\\red255\\green110\\blue180;\n\\red238\\green106\\blue167;\n\\red205\\green96\\blue144;\n\\red139\\green58\\blue98;\n\\red205\\green92\\blue92;\n\\red255\\green106\\blue106;\n\\red238\\green99\\blue99;\n\\red205\\green85\\blue85;\n\\red139\\green58\\blue58;\n\\red255\\green255\\blue240;\n\\red255\\green255\\blue240;\n\\red238\\green238\\blue224;\n\\red205\\green205\\blue193;\n\\red139\\green139\\blue131;\n\\red240\\green230\\blue140;\n\\red255\\green246\\blue143;\n\\red238\\green230\\blue133;\n\\red205\\green198\\blue115;\n\\red139\\green134\\blue78;\n\\red230\\green230\\blue250;\n\\red255\\green240\\blue245;\n\\red255\\green240\\blue245;\n\\red238\\green224\\blue229;\n\\red205\\green193\\blue197;\n\\red139\\green131\\blue134;\n\\red124\\green252\\blue0;\n\\red255\\green250\\blue205;\n\\red255\\green250\\blue205;\n\\red238\\green233\\blue191;\n\\red205\\green201\\blue165;\n\\red139\\green137\\blue112;\n\\red173\\green216\\blue230;\n\\red191\\green239\\blue255;\n\\red178\\green223\\blue238;\n\\red154\\green192\\blue205;\n\\red104\\green131\\blue139;\n\\red240\\green128\\blue128;\n\\red224\\green255\\blue255;\n\\red224\\green255\\blue255;\n\\red209\\green238\\blue238;\n\\red180\\green205\\blue205;\n\\red122\\green139\\blue139;\n\\red238\\green221\\blue130;\n\\red255\\green236\\blue139;\n\\red238\\green220\\blue130;\n\\red205\\green190\\blue112;\n\\red139\\green129\\blue76;\n\\red250\\green250\\blue210;\n\\red211\\green211\\blue211;\n\\red144\\green238\\blue144;\n\\red211\\green211\\blue211;\n\\red255\\green182\\blue193;\n\\red255\\green174\\blue185;\n\\red238\\green162\\blue173;\n\\red205\\green140\\blue149;\n\\red139\\green95\\blue101;\n\\red255\\green160\\blue122;\n\\red255\\green160\\blue122;\n\\red238\\green149\\blue114;\n\\red205\\green129\\blue98;\n\\red139\\green87\\blue66;\n\\red32\\green178\\blue170;\n\\red135\\green206\\blue250;\n\\red176\\green226\\blue255;\n\\red164\\green211\\blue238;\n\\red141\\green182\\blue205;\n\\red96\\green123\\blue139;\n\\red132\\green112\\blue255;\n\\red119\\green136\\blue153;\n\\red119\\green136\\blue153;\n\\red176\\green196\\blue222;\n\\red202\\green225\\blue255;\n\\red188\\green210\\blue238;\n\\red162\\green181\\blue205;\n\\red110\\green123\\blue139;\n\\red255\\green255\\blue224;\n\\red255\\green255\\blue224;\n\\red238\\green238\\blue209;\n\\red205\\green205\\blue180;\n\\red139\\green139\\blue122;\n\\red50\\green205\\blue50;\n\\red250\\green240\\blue230;\n\\red255\\green0\\blue255;\n\\red255\\green0\\blue255;\n\\red238\\green0\\blue238;\n\\red205\\green0\\blue205;\n\\red139\\green0\\blue139;\n\\red176\\green48\\blue96;\n\\red255\\green52\\blue179;\n\\red238\\green48\\blue167;\n\\red205\\green41\\blue144;\n\\red139\\green28\\blue98;\n\\red102\\green205\\blue170;\n\\red0\\green0\\blue205;\n\\red186\\green85\\blue211;\n\\red224\\green102\\blue255;\n\\red209\\green95\\blue238;\n\\red180\\green82\\blue205;\n\\red122\\green55\\blue139;\n\\red147\\green112\\blue219;\n\\red171\\green130\\blue255;\n\\red159\\green121\\blue238;\n\\red137\\green104\\blue205;\n\\red93\\green71\\blue139;\n\\red60\\green179\\blue113;\n\\red123\\green104\\blue238;\n\\red0\\green250\\blue154;\n\\red72\\green209\\blue204;\n\\red199\\green21\\blue133;\n\\red25\\green25\\blue112;\n\\red245\\green255\\blue250;\n\\red255\\green228\\blue225;\n\\red255\\green228\\blue225;\n\\red238\\green213\\blue210;\n\\red205\\green183\\blue181;\n\\red139\\green125\\blue123;\n\\red255\\green228\\blue181;\n\\red255\\green222\\blue173;\n\\red255\\green222\\blue173;\n\\red238\\green207\\blue161;\n\\red205\\green179\\blue139;\n\\red139\\green121\\blue94;\n\\red0\\green0\\blue128;\n\\red0\\green0\\blue128;\n\\red253\\green245\\blue230;\n\\red107\\green142\\blue35;\n\\red192\\green255\\blue62;\n\\red179\\green238\\blue58;\n\\red154\\green205\\blue50;\n\\red105\\green139\\blue34;\n\\red255\\green165\\blue0;\n\\red255\\green165\\blue0;\n\\red238\\green154\\blue0;\n\\red205\\green133\\blue0;\n\\red139\\green90\\blue0;\n\\red255\\green69\\blue0;\n\\red255\\green69\\blue0;\n\\red238\\green64\\blue0;\n\\red205\\green55\\blue0;\n\\red139\\green37\\blue0;\n\\red218\\green112\\blue214;\n\\red255\\green131\\blue250;\n\\red238\\green122\\blue233;\n\\red205\\green105\\blue201;\n\\red139\\green71\\blue137;\n\\red238\\green232\\blue170;\n\\red152\\green251\\blue152;\n\\red154\\green255\\blue154;\n\\red144\\green238\\blue144;\n\\red124\\green205\\blue124;\n\\red84\\green139\\blue84;\n\\red175\\green238\\blue238;\n\\red187\\green255\\blue255;\n\\red174\\green238\\blue238;\n\\red150\\green205\\blue205;\n\\red102\\green139\\blue139;\n\\red219\\green112\\blue147;\n\\red255\\green130\\blue171;\n\\red238\\green121\\blue159;\n\\red205\\green104\\blue137;\n\\red139\\green71\\blue93;\n\\red255\\green239\\blue213;\n\\red255\\green218\\blue185;\n\\red255\\green218\\blue185;\n\\red238\\green203\\blue173;\n\\red205\\green175\\blue149;\n\\red139\\green119\\blue101;\n\\red205\\green133\\blue63;\n\\red255\\green192\\blue203;\n\\red255\\green181\\blue197;\n\\red238\\green169\\blue184;\n\\red205\\green145\\blue158;\n\\red139\\green99\\blue108;\n\\red221\\green160\\blue221;\n\\red255\\green187\\blue255;\n\\red238\\green174\\blue238;\n\\red205\\green150\\blue205;\n\\red139\\green102\\blue139;\n\\red176\\green224\\blue230;\n\\red160\\green32\\blue240;\n\\red155\\green48\\blue255;\n\\red145\\green44\\blue238;\n\\red125\\green38\\blue205;\n\\red85\\green26\\blue139;\n\\red255\\green0\\blue0;\n\\red255\\green0\\blue0;\n\\red238\\green0\\blue0;\n\\red205\\green0\\blue0;\n\\red139\\green0\\blue0;\n\\red188\\green143\\blue143;\n\\red255\\green193\\blue193;\n\\red238\\green180\\blue180;\n\\red205\\green155\\blue155;\n\\red139\\green105\\blue105;\n\\red65\\green105\\blue225;\n\\red72\\green118\\blue255;\n\\red67\\green110\\blue238;\n\\red58\\green95\\blue205;\n\\red39\\green64\\blue139;\n\\red139\\green69\\blue19;\n\\red250\\green128\\blue114;\n\\red255\\green140\\blue105;\n\\red238\\green130\\blue98;\n\\red205\\green112\\blue84;\n\\red139\\green76\\blue57;\n\\red244\\green164\\blue96;\n\\red46\\green139\\blue87;\n\\red84\\green255\\blue159;\n\\red78\\green238\\blue148;\n\\red67\\green205\\blue128;\n\\red46\\green139\\blue87;\n\\red255\\green245\\blue238;\n\\red255\\green245\\blue238;\n\\red238\\green229\\blue222;\n\\red205\\green197\\blue191;\n\\red139\\green134\\blue130;\n\\red160\\green82\\blue45;\n\\red255\\green130\\blue71;\n\\red238\\green121\\blue66;\n\\red205\\green104\\blue57;\n\\red139\\green71\\blue38;\n\\red135\\green206\\blue235;\n\\red135\\green206\\blue255;\n\\red126\\green192\\blue238;\n\\red108\\green166\\blue205;\n\\red74\\green112\\blue139;\n\\red106\\green90\\blue205;\n\\red131\\green111\\blue255;\n\\red122\\green103\\blue238;\n\\red105\\green89\\blue205;\n\\red71\\green60\\blue139;\n\\red112\\green128\\blue144;\n\\red198\\green226\\blue255;\n\\red185\\green211\\blue238;\n\\red159\\green182\\blue205;\n\\red108\\green123\\blue139;\n\\red112\\green128\\blue144;\n\\red255\\green250\\blue250;\n\\red255\\green250\\blue250;\n\\red238\\green233\\blue233;\n\\red205\\green201\\blue201;\n\\red139\\green137\\blue137;\n\\red0\\green255\\blue127;\n\\red0\\green255\\blue127;\n\\red0\\green238\\blue118;\n\\red0\\green205\\blue102;\n\\red0\\green139\\blue69;\n\\red70\\green130\\blue180;\n\\red99\\green184\\blue255;\n\\red92\\green172\\blue238;\n\\red79\\green148\\blue205;\n\\red54\\green100\\blue139;\n\\red210\\green180\\blue140;\n\\red255\\green165\\blue79;\n\\red238\\green154\\blue73;\n\\red205\\green133\\blue63;\n\\red139\\green90\\blue43;\n\\red216\\green191\\blue216;\n\\red255\\green225\\blue255;\n\\red238\\green210\\blue238;\n\\red205\\green181\\blue205;\n\\red139\\green123\\blue139;\n\\red255\\green99\\blue71;\n\\red255\\green99\\blue71;\n\\red238\\green92\\blue66;\n\\red205\\green79\\blue57;\n\\red139\\green54\\blue38;\n\\red64\\green224\\blue208;\n\\red0\\green245\\blue255;\n\\red0\\green229\\blue238;\n\\red0\\green197\\blue205;\n\\red0\\green134\\blue139;\n\\red238\\green130\\blue238;\n\\red208\\green32\\blue144;\n\\red255\\green62\\blue150;\n\\red238\\green58\\blue140;\n\\red205\\green50\\blue120;\n\\red139\\green34\\blue82;\n\\red245\\green222\\blue179;\n\\red255\\green231\\blue186;\n\\red238\\green216\\blue174;\n\\red205\\green186\\blue150;\n\\red139\\green126\\blue102;\n\\red245\\green245\\blue245;\n\\red255\\green255\\blue0;\n\\red255\\green255\\blue0;\n\\red238\\green238\\blue0;\n\\red205\\green205\\blue0;\n\\red139\\green139\\blue0;\n\\red154\\green205\\blue50;\n} \n{\\stylesheet{\\s1 heading 1;}\n{\\s2 heading 2;}\n{\\s3 heading 3;}\n{\\s4 heading 4;}\n}\n\n"
#> 
#> $body
#> [1] "{\\header\n{\\pard\\tqr\\tx12960\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b  \\tab Page \\chpgn  of {\\field{\\*\\fldinst NUMPAGES }} }\\par}\n{\\pard\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b }\\par}\n}\n{\\footer\n{\\pard\\tqr\\tx12960\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b DB lock (snapshot): Analysis 2025 \\tab Created on: 14APR2026 00:24:34 UTC}\\par}\n{\\pard\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b Program name: C:\\\\program\\\\path\\\\code.r}\\par}\n{\\pard\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b Output name: C:\\\\output\\\\path\\\\MyColSpanTable.rtf}\\par}\n}\n\\paperw15840\\paperh12240\\landscape\n\n\\margl1440\\margr1440\\margt1080\\margb1080\\headery720\\footery720\n\n{\\pard\\hyphpar\\sb180\\sa180\\fi0\\li0\\ri0\\ql\\fs24\\s2{\\f5\\b Table 1. My table}\\par}\n\n\n{\\*\\bkmkstart MyColSpanTable}\\trowd\\trgaph108\\trleft0\\trqc\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx3535\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx8248\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrr\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx12961\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 }\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 Cspan1}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 CSpan2}\\cell\n\\intbl\\row\\pard\n\\trowd\\trgaph108\\trleft0\\trqc\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx3535\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx5891\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx8247\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx10603\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrr\\brdrs\\brdrw15\\clbrdrb\\brdrw15\\clcbpat247\\clvertalb\\cellx12959\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 Col1}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 Col2}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 Col3}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 Col4}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5\\b\\cf24\\chshdng0\\chcbpat247\\cb247 Col5}\\cell\n\\intbl\\row\\pard\n\\trowd\\trgaph108\\trleft0\\trqc\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrs\\brdrw15\\clvertalt\\cellx3535\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrs\\brdrw15\\clvertalt\\cellx5891\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrs\\brdrw15\\clvertalt\\cellx8247\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrb\\brdrs\\brdrw15\\clvertalt\\cellx10603\n\\clbrdrl\\brdrs\\brdrw15\\clbrdrt\\brdrs\\brdrw15\\clbrdrr\\brdrs\\brdrw15\\clbrdrb\\brdrs\\brdrw15\\clvertalt\\cellx12959\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\ql\\fs20{\\f5 1}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5 2}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5 3}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5 4}\\cell\n\\pard\\hyphpar0\\sb15\\sa15\\fi0\\li0\\ri0\\qc\\fs20{\\f5 5}\\cell\n\\intbl\\row\\pard\n\n\n"
#> 
#> $end
#> [1] "{\\*\\bkmkend MyColSpanTable}\n}"
#>